Refactored generator logic so there wasn't as much duplication.
