+++ lib/enum_weasel.rb
@@ -1,8 +1,8 @@
 require "active_record"
-require "active_support/all"
 require "enum_weasel/version"
 require "enum_weasel/get_all_model_enums"
 require "enum_weasel/generators"
+require "enum_weasel/generators/tables/migration"
 require "enum_weasel/generators/tables/creation"
 require "enum_weasel/generators/tables/deletion"
 require "enum_weasel/generators/data/creation"
